pyproject.toml
[tool.ruff]
line-length = 120
indent-width = 4

[tool.ruff.lint]
select = ["E", # pycodestyle errors
    "W", # pycodestyle warnings
    "F", # pyflakes
    "I", # isort
    "C", # flake8-comprehensions
    "B", # flake8-bugbear
    "N",
    "DJ",
    "ICN",
    "SIM",
    "PL",
]
ignore = ["F405", "F403", "D103", "D100", "D102", "D101", "ANN001", "ANN003", "ANN201", "ERA001"]

[tool.ruff.lint.per-file-ignores]
"tests/*" = ["S101", "S105", "S106", "PLR2004"]
"**/migrations/*" = ["E501"]

[tool.poetry]
name = "registry"
version = "0.1.0"
description = ""
authors = ["NickolayBabulich <nickolas.grn@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.13"
django = "^5.1.2"
django-environ = "^0.11.2"
psycopg2-binary = "^2.9.10"
xlsxwriter = "^3.2.0"
reportlab = "^4.2.5"
djangorestframework = "^3.15.2"
django-filter = "^24.3"
drf-spectacular = "^0.28.0"
pandas = "^2.2.3"
openpyxl = "^3.1.5"


[tool.poetry.group.dev.dependencies]
ruff = "^0.7.2"
django-debug-toolbar = "^4.4.6"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
